+++ CHANGELOG.md
@@ -22,8 +22,10 @@ HumHub Changelog
 - Enh #7753: New modal full screen size
 - Fix #7754: Tour module - the bottom part of the Profile image is hidden
 - Fix #7752: Fix content resetForm after form submit
+- Fix #7759: Flush content search cache after update a record
 - Fix #7760: Content form menu text color same as background
 - Fix #7764: File Handler Dropdown menu items styling
+- Fix #7767: Added PK in `content_fulltext` for Percona-XtraDB-Cluster installations
 - Enh: Implements Button `disabled()`
 
 1.18.0-beta.4 (September 19, 2025)